requirements.txt
pytest==8.1.1
pytest-asyncio==0.18.3
appium-python-client==4.0.0
selenium==4.19.0
